refactor(sdk): unify pending-reply bookkeeping into pending-replies.ts
Move the four data-channel request/response disciplines (type-keyed
reply slots, id-correlated JSON-RPC calls, wake_up/goto_sleep motion
completions, predicate-matched broadcast waiters) into one
PendingReplies ledger with a single settleAll() used by every
session-teardown path. Sending stays with the class (it owns the dc);
the ledger owns the waiters only. Also fixes two teardown gaps the
per-mechanism calls had left: broadcast waiters were never settled
(they waited out their own timeout on a dead channel) and disconnect()
never rejected pending rpcCalls.
